added a function to save the processed subject in the fibrosis project
to avoid reprocessing.

+++ script/mouse_data_preparation.py
@@ -10,26 +10,37 @@ import shutil
 def mouse_fibrosis_data_preparation(input_data, root_path, work_dir, crop=False, clean=False):
 
     raw_data, mask_paths = batch_processing(input_data, root=root_path)
-    
+    processed_subs = []
+    if os.path.isfile(os.path.join(work_dir, 'processed_subjects.txt')):
+        with open(os.path.join(work_dir, 'processed_subjects.txt'), 'r') as f:
+            for line in f:
+                processed_subs.append(line.strip())
+
     for i, raw_data_folder in enumerate(raw_data):
-        filename, folder_name = mouse_lung_data_preparation(raw_data_folder, work_dir)
-        
-        converter = DicomConverters(filename, clean=clean)
-        converted_data = converter.mitk_converter()
-
-        for j, mask in enumerate(os.listdir(mask_paths[i])):
-            new_folder = folder_name+'/mouse_{}'.format(str(j+1).zfill(2))
-            os.mkdir(new_folder)
-            prefix = 'Raw_data_for_{}'.format(mask.split('.')[0])
-            if crop:
-                image, mask = cropping(converted_data, os.path.join(mask_paths[i], mask), prefix=prefix)
-                shutil.move(image, new_folder)
-                shutil.move(mask, new_folder)
-            else:
-                shutil.copy2(os.path.join(mask_paths[i], mask), new_folder)
-                shutil.copy2(converted_data, os.path.join(new_folder, prefix))
-        os.remove(converted_data)
-
+        if raw_data_folder not in processed_subs:
+            filename, folder_name = mouse_lung_data_preparation(raw_data_folder, work_dir)
+    
+            converter = DicomConverters(filename, clean=clean)
+            converted_data = converter.mitk_converter()
+    
+            for j, mask in enumerate(os.listdir(mask_paths[i])):
+                new_folder = folder_name+'/mouse_{}'.format(str(j+1).zfill(2))
+                if not os.path.isdir(new_folder):
+                    os.mkdir(new_folder)
+                prefix = 'Raw_data_for_{}'.format(mask.split('.')[0])
+                if crop:
+                    image, mask = cropping(converted_data, os.path.join(mask_paths[i], mask),
+                                           prefix=prefix)
+                    shutil.move(image, new_folder)
+                    shutil.move(mask, new_folder)
+                else:
+                    shutil.copy2(os.path.join(mask_paths[i], mask), new_folder)
+                    shutil.copy2(converted_data, os.path.join(new_folder, prefix))
+            os.remove(converted_data)
+            with open(os.path.join(work_dir, 'processed_subjects.txt'), 'a') as f:
+                f.write(raw_data_folder+'\n')
+        else:
+            print('{} has already been processed and will be skipped.'.format(raw_data_folder))
 
 if __name__ == "__main__":
